supabase_listener/requirements.txt
requests
websocket-client
flask
APScheduler
supabase
gunicorn
asyncio